Vehicle Tracker/requirements.txt
opencv-python~=4.6.0.66
depthai==2.16.0.0
blobconverter==1.2.9

numpy~=1.23.4
mysql-connector-python~=8.0.31